requirements.txt
Django==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.11.0